add recursive check for children when parent property is missing of type 'object' and of type 'array' with 'items'
if a property has children (child schemas) and this property is missing then the children would not get checked.  If the children, for example, are required those errors would never show up. This makes sure that all errors on all child properties are returned. In most case the parent is just a holder for the child properties and if it's missing it's not that interesting to know that it's missing but if its children are missing, which may correspond for form fields, it would be very useful to know since the focus was initially on the child fields.